refactor: separate un api by a module
